GBC Elevator V1.0 - Lego Digital Designer
I spent a couple hours working on a very rough version of the elevator
idea I've had in my head. It uses blocks as a wall to keep the balls
inside the elevator, then it rotates away from the elevator to let the
balls out. I haven't worked out the system that would move the elevator
up and down yet, but I'm sure someone could work it out after some time.