Improved Bicept and Sensor Mount
So these are just the improved versions of our old models, it now has a
base plate (next class we will switch out the green bottom plates for
regular base plates because they are too thick), it now has a emergency
stopper bar just incase the motor pushes it too far, and we added the
basket(s) where the new bucket at the top of the arm will dump the
balls into. Also, we agreed that our model will connect with Bomba
Phonexes' model.